Solucionados Bugs de inclusion y punto de entrada

+++ Pprueba/Src/PhysicsManager.cpp
@@ -45,7 +45,7 @@ void PhysicsManager::stepUp(double deltaTime/*seconds*/)
 {
 	//deltaTime = 0.0016;
 	unsigned char maxSubSteps = 10;	//esto hay que calcularlo de manera dinámica, pero no sé como
-	world->stepSimulation(deltaTime, maxSubSteps);
+	world->stepSimulation((float)deltaTime, maxSubSteps);
 }
 
 void PhysicsManager::addRigidBody(btRigidBody * newRigidBody)
